completed multiply instructions and added fucntions to shift bits from structs to the right place

+++ src/assembler/register_to_number.h
@@ -2,5 +2,5 @@
 #define REGISTER_TO_NUMBER_H
 #include <stdint.h>
 char* remove_non_integer_characters(char *input);
-uint8_t register_to_number(char *reg); 
+int register_to_number(char *reg); 
 #endif // REGISTER_TO_NUMBER_H
